Update conn_pool.go
Declared ConnPoolTimeout as an exportable variable. This will enable the programs that use this client to set it to a value that is required. Programs that use this can also perform some actions upon getting the timeout. Proposed workaround for the problem specified [here](http://cbugg.hq.couchbase.com/bug/bug-774).

Whenever, HttpClient is used to make requests, timeout can be specified by setting HttpClient to a custom http.Client as mentioned [here](http://stackoverflow.com/a/16895878/1043536)